--- a/spec/features/users/form_answers/past_applications_spec.rb
+++ b/spec/features/users/form_answers/past_applications_spec.rb
@@ -65,10 +65,8 @@ So that I see
         visit dashboard_path
       end
 
-      it "should display past successful applications" do
-        save_and_open_page
-
-        # TODO: need to finish this spec
+      xit "should display past successful applications" do
+        # save_and_open_page
       end
     end