--- a/spec/rails_helper.rb
+++ b/spec/rails_helper.rb
@@ -1,28 +1,32 @@
-ENV["RAILS_ENV"] ||= 'test'
+ENV["RAILS_ENV"] ||= "test"
 ENV["VIRUS_SCANNER_API_URL"] ||= "http://virus.scanner"
 ENV["VIRUS_SCANNER_API_KEY"] ||= "random_api_key"
 
-require 'simplecov'
+require "simplecov"
 require "codeclimate-test-reporter"
-SimpleCov.add_filter 'vendor'
+SimpleCov.add_filter "vendor"
 SimpleCov.formatters = []
 SimpleCov.start CodeClimate::TestReporter.configuration.profile
 
+require "spec_helper"
 require File.expand_path("../../config/environment", __FILE__)
-require 'rspec/rails'
-require 'capybara/rspec'
-require 'shoulda/matchers'
+require "rspec/rails"
+require "capybara/rspec"
+require "shoulda/matchers"
 require "capybara-screenshot/rspec"
 require "webmock/rspec"
 
-Capybara.javascript_driver = :webkit
+WebMock.disable_net_connect!(allow: "codeclimate.com", allow_localhost: true)
+
+# Require all support files.
 Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
+
+Capybara.javascript_driver = :webkit
+Capybara::Screenshot.webkit_options = { width: 1024, height: 768 }
 
 ActiveRecord::Migration.check_pending!
 ActiveRecord::Migration.maintain_test_schema!
 Qae::Application.load_tasks
-Capybara::Screenshot.webkit_options = { width: 1024, height: 768 }
-WebMock.disable_net_connect!(allow: /codeclimate.com/, allow_localhost: true)
 
 RSpec.configure do |config|
   config.include FactoryGirl::Syntax::Methods
@@ -30,13 +34,9 @@ RSpec.configure do |config|
   config.include ExpectationHelper, type: :feature
   config.include Devise::TestHelpers, type: :controller
 
-  config.filter_run focus: true
-  config.run_all_when_everything_filtered = true
-
   config.raise_error_for_unimplemented_steps = true
   config.use_transactional_fixtures = false
   config.infer_base_class_for_anonymous_controllers = false
-  config.order = "random"
 
   config.before :each do
     stub_request(:post, /virus.scanner/).